fixed look of chat container
