+++ src/components/pages/EventsPage/ViewEvents/SingleEvent/ViewSingleEvent.tsx
@@ -1,12 +1,11 @@
 import { useGetSingleEvent, useUpdateEvent } from "@/api/events";
 import styles from "../EventStyles.module.css";
 import { useParams } from "react-router-dom";
-import SuggestionsContainer from "../SuggestionsContainer";
-import SuggestionForm from "./SuggestionForm";
+import SuggestionsContainer from "./Suggestions/SuggestionsContainer";
+import SuggestionForm from "./Suggestions/SuggestionForm";
 import { useAuth } from "@/context/auth/auth";
 import { Switch } from "@/components/ui/switch";
 import Chat from "./Chat/Chat";
-
 
 export default function ViewSingleEvent() {
   const { user } = useAuth();
@@ -37,7 +36,6 @@ export default function ViewSingleEvent() {
 
   return (
     <div className={styles.eventContainer}>
-
       <div className="col1">
         {/* ---Event Name----*/}
         <div>
@@ -45,27 +43,28 @@ export default function ViewSingleEvent() {
           <p>{event.time}</p>
 
           {/* ---Description----*/}
-          <p style={{ marginTop: "2em" }}>
-            {event.description}
-          </p>
+          <p style={{ marginTop: "2em" }}>{event.description}</p>
         </div>
 
         <div style={{ marginTop: "3em" }}>
-
           {/* ----Admin Toggle Switches---- */}
           {/* only admin creator can toggle switches */}
           {user._id === event.ownerId && (
             <div>
-
               {/* suggestions switch */}
-              <div className="flex items-center space-x-2 mb-4"
-                style={{ marginBottom: "3em" }}>
+              <div
+                className="flex items-center space-x-2 mb-4"
+                style={{ marginBottom: "3em" }}
+              >
                 <Switch
                   checked={event.suggestionsEnabled}
                   onCheckedChange={toggleSuggestions}
                   id="toggle-suggestions"
                 />
-                <label htmlFor="toggle-suggestions" className="text-sm font-medium">
+                <label
+                  htmlFor="toggle-suggestions"
+                  className="text-sm font-medium"
+                >
                   {event.suggestionsEnabled
                     ? "Disable Suggestions"
                     : "Enable Suggestions"}
@@ -73,39 +72,38 @@ export default function ViewSingleEvent() {
               </div>
 
               {/* chat switch */}
-              <div className="flex items-center space-x-2 mb-4"
-                style={{ marginBottom: "3em" }}>
+              <div
+                className="flex items-center space-x-2 mb-4"
+                style={{ marginBottom: "3em" }}
+              >
                 <Switch
                   checked={event.chatEnabled}
                   onCheckedChange={toggleChat}
                   id="toggle-suggestions"
                 />
-                <label htmlFor="toggle-suggestions" className="text-sm font-medium">
-                  {event.chatEnabled
-                    ? "Disable Chat"
-                    : "Enable Chat"}
+                <label
+                  htmlFor="toggle-suggestions"
+                  className="text-sm font-medium"
+                >
+                  {event.chatEnabled ? "Disable Chat" : "Enable Chat"}
                 </label>
               </div>
             </div>
           )}
 
+          {/* ----Suggestions Container--- */}
+          {event.suggestionsEnabled && (
+            <>
+              <SuggestionForm event={event} />{" "}
+              {/* Add A Suggestion button lives here*/}
+              <SuggestionsContainer event={event} />
+            </>
+          )}
 
           {/* ----Chat Container---- */}
           {event.chatEnabled && <Chat eventId={eventId} />}
-
-          {/* ----Suggestions Container--- */}
-          {event.suggestionsEnabled &&
-            <>
-              <SuggestionForm event={event} /> {/* Add A Suggestion button lives here*/}
-              <SuggestionsContainer event={event} />
-            </>
-          }
-
         </div>
       </div>
-
-
-
     </div>
   );
 }
